+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 **General**:
 - Fixed an issue that could cause card images to not load.
 
+**Battlegrounds**:
+- Fixed the hero guide not loading when starting HDT mid-match.
+
 ## **Release v1.54.1 - 2026-07-30**
 **Hearthstone**:
 - Added an option to show the order cards entered the board, numbering each minion and weapon.\